necesario

+++ CEPROZAC/database/migrations/2018_04_23_201749_EntradasAgroquimicosMigration.php
@@ -20,7 +20,7 @@ class EntradasAgroquimicosMigration extends Migration
             $table->string('factura');
             $table->integer('id_material')->unsigned();
             $table->foreign('id_material')->references('id')->on('almacenagroquimicos');
-            $table->integer('cantidad');
+            $table->double('cantidad');
             $table->string('medida');
             $table->string('medidaaux');
             $table->double('p_unitario');
